requirements.txt
fastf1
pandas
scikit-learn
numpy
matplotlib
Pillow
